# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Nopic web application's API",
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "params",
                        "in": "body",
                        "description": "Parameters to pass (in body)",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/loginCmsct": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "",
                "operationId": "loginCmsct",
                "parameters": [
                    {
                        "name": "params",
                        "in": "body",
                        "description": "Parameters to pass (in body)",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/registerCmsct": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "",
                "operationId": "registerCmsct",
                "parameters": [
                    {
                        "name": "params",
                        "in": "body",
                        "description": "Parameters to pass (in body)",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "params",
                        "in": "body",
                        "description": "Parameters to pass (in body)",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "",
                "operationId": "userTest",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Use the keyword 'Bearer' before the access-key. Example 'Bearer 0123456789_access-key-goes-here_987654321",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/cmsct": {
            "get": {
                "tags": [
                    "cmsct"
                ],
                "summary": "",
                "operationId": "cmsctTest",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Use the keyword 'Bearer' before the access-key. Example 'Bearer 0123456789_access-key-goes-here_987654321",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "definitions": {}
}